updated version with 'Categories' navigation

+++ coach-bookings.php
@@ -171,11 +171,17 @@ if (mysqli_num_rows($result) == 1) {
     
     <script>
       function coach_change_payment_status(bcch_id){
-        location.href="coach-change-payment-status?id="+bcch_id;
+        var ok = confirm("Do you want to change the payment status of the booker?")
+        if (ok) {
+          location.href="coach-change-payment-status?id="+bcch_id;
+        }
       }
 
       function coach_delete_book(bcch_id){
-        location.href="coach-delete-book?id="+bcch_id;
+        var ok = confirm("Do you want to delete this particular booking?")
+        if (ok) {
+          location.href="coach-delete-book?id="+bcch_id;
+        }
       }
 
       let sidebar = document.querySelector(".sidebar");
